# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Lunar Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:80",
    "basePath": "/",
    "paths": {
        "/api/missions": {
            "get": {
                "description": "Возвращает все миссии с фильтрацией по статусу и дате формирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Миссии"
                ],
                "summary": "Получить все миссии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус перевозки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "начальная дата формирования",
                        "name": "formation_date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "конечная дата формирвания",
                        "name": "formation_completion_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllMissionsResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Позволяет изменить название миссии и возвращает обновлённые данные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Миссии"
                ],
                "summary": "Указать название",
                "parameters": [
                    {
                        "description": "Название",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SwaggerUpdateMissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Удаляет черновую миссию",
                "tags": [
                    "Миссии"
                ],
                "summary": "Удалить черновую миссию",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/missions/delete_module/{id}": {
            "delete": {
                "description": "Удалить модуль из черновой миссии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Миссии"
                ],
                "summary": "Удалить модуль из черновой миссии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id модуля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/missions/{id}": {
            "get": {
                "description": "Возвращает подробную информацию о миссии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Миссии"
                ],
                "summary": "Получить одну миссию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id миссии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.MissionResponse"
                        }
                    }
                }
            }
        },
        "/api/missions/{id}/moderator_confirm": {
            "put": {
                "description": "Подтвердить или отменить миссию модератором",
                "tags": [
                    "Миссии"
                ],
                "summary": "Подтвердить миссию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id миссии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.SwaggerUpdateMissionRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "ds.Module": {
            "type": "object",
            "required": [
                "description",
                "diameter",
                "length",
                "mass"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "diameter": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "length": {
                    "type": "string"
                },
                "mass": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.AllMissionsResponse": {
            "type": "object",
            "properties": {
                "missions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemes.MissionOutput"
                    }
                }
            }
        },
...
        "schemes.MissionOutput": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.MissionResponse": {
            "type": "object",
            "properties": {
                "mission": {
                    "$ref": "#/definitions/schemes.MissionOutput"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Module"
                    }
                }
            }
        },
...
    }
}